Add visitor pattern and encoding/decoding API
To expose the encoded/decoded configuration (i.e. bit-formatting for the
hardware), each container should implement three classes of functions:

- Both `read_addresses()` and `write_addresses()` or a single
  `addresses()` member function.  These functions return the addresses
  in memory the generated configuration words should be read from or
  written to.  In case the addresses for reading and writing of the
  container are the same, a single function `addresses()` is preferred.
  These functions should take the global coordinate of the container as
  their only argument and return an array of addresses, e.g.:

  std::array<uint32_t, 1> CapMemCell::addresses(CapMemCellOnDLS const&) const;

- `encode()`, which should return an array of configuration words of
  the same length as the number of addresses.  This function can
  optionally accept the global coordinate of the container, e.g.:

  std::array<uint32_t, 1> CapMemCell::encode() const;  // or:
  std::array<uint32_t, 1> CapMemCell::encode(CapMemCellOnDLS const&) const;

- `decode()`, which should accept an array of configuration words
  or the global coordinate of the container and an array of
  configuration words, e.g.:

  void CapMemCell::decode(std::array<uint32_t, 1> const&);  // or:
  void CapMemCell::decode(CapMemCellOnDLS const&, std::array<uint32_t, 1> const&);

Note that the sole responsibility of these functions is to generate
configuration instructions local to the respective container, i.e.
`CapMem` is not responsible for recursively calling the `address`,
`encode` or `decode` member functions of `CapMemCell`, even though it
holds an array of `CapMemCell`s [1].
As in this case all configuration bits are contained in `CapMemCell` and
no other local member variables remain in `CapMem`, it is not necessary
to implement the `address`, `encode` or `decode` member functions.
Instead, `CapMem` is tagged with `typedef std::false_type has_local_data`.

[1]: The newly introduced `visit_preorder` function is used to do a
pre-order tree traversal of the container hierarchy, ensuring that the
corresponding member functions are called for each container and in the
same order.

Change-Id: I278ceb79fdfbb1a4fcf953bc3dd66f9de921809a