Change default CapMem counter speed
There exists a problem with the CapMem which prevents it from
driving very small currents. The bug is known but is still present
in Hicann-X, it got even more severe. The CapMem starts to drive
currents only at a setting of roughly 50 LSB with the previous default
values (while some chip quadrants start even close to 0).
The issue is reported at
https://brainscales-r.kip.uni-heidelberg.de/projects/hicann-dls/work_packages/2403/

Reducing the speed of the CapMem ramp has made the problem a little
less severe, with most quadrants starting to drive currents between
10...35 LSB. This changeset sets the counter speed to half the previous
values. Combined with a lower current onto the ramp capacitor this
results in a working configuration that enables lower currents.

Change-Id: Ia3ad77f80fc70dc01f0110b82e337fb678879b03