Ignore possible doubly defined option
Depends-On: 9548

`--with-munge` determines if `waf` should look for `libmunge` during
configuration. It is currently only used by `haldls` but will, in the
future, also used in `fisch`.

It makes no sense to split the options because either we want to look
for munge or we do not.

Hence, we just ignore the resulting ArgumentError if `fisch` was
recursed first.

Since only one of `haldls` or `fisch` can define the option, move it
from the `haldls` option group to the general one.

Change-Id: Ic3889342d55d24ce8672629c9df723b72ab06598