Add default settings to neuron backend config
Add config to neuron backend which allow normal usage.

* The reset holdoff is only turned off when a value of 15 is configured.
Therefore, I set the default to 15 instead of 0. (This is already stated
in the docstring of the respective setting).

* When not using bayesian extensions, enable_neuron_master still has to
be turned on in order to send postpulses to the synapses. Working
towards issue 3368, setting a value of true is a more intuitive default.

Change-Id: I823fe6228ef741c0cbaab78d462d32500d02761e