Remove pureness from PlaybackGenerator destructor
A pure virtual destructor is not needed since `PlaybackGenerator`
already has another pure virtual function.

Note: This change was initiated due to linking errors: several object
files for `playback_generator.cpp` with the same name were generated
and then linked incorrectltly. Moving the impl. of the destructor to the
header file allows us to delete `playback_generator.cpp`.
Removing pureness is **not** needed in order to move the impl. in the
header but saves us some lines of codes.

Change-Id: I9826a858a733cccf733c6163e739f8826a4a6025