Fixes for bayesian computation in neuron backend
* Add more documentation
* Swap wrongly assigned bayesian mode bits
* Create enum-type operation selection in lola
* Fix spelling of baYesian within haldls implementation (although the
  `y` is missing in the schematic.)

Also, we change the default configuration of the digital neuron backend:

Previously, we have used the `neuron_fire_connect` signal to trigger
post pulses (which happens bayesian extensions are disabled).
This signal is generated by a master neuron (hence we enabled
`neuron_master`).

Now, the default is using the bayesian extensions to generate the post
pulse, and using only the local refractory state there. It would not
hurt to enable `neuron_master` now, but we don't need it.

This new default does not change the behaviour of the circuitry.
But it makes more sense, as the `neuron_fire_connect` signal is
used to connect neurons, with the master/slave settings determining
who controls the refractory state, but we don't need that for
individual, disconnected neurons.

Change-Id: Ic957ef73d96755afd6a8eebc6874fdb4e3028bb8