fix(pickle): use new-style pickle __setstate__
I wasn't able to fix the placement new method completely: The check for
holder_constructed() helped to fix the `inited_obj.__setstate__(state)`
case, but there was still a leak in the
`only_newed_obj.__setstate__(state)` case (which is what the pickle
library does). Using the public API of pybind11::object/handle also
proved difficult; now we resort to using the internal pybind11 pickle
API. We construct the C++ state, pass it to pybind11's "setstate"
function and use it as is.

Change-Id: I74bb3645956c0260cd80f935c90009cf6d8b1948